fix: implemented header component in osc through building blobk common component
