+++ apps/retail/components/header/index.tsx
@@ -30,7 +30,7 @@ const Header = () => {
     <Box>
       {renderTopHeader && (
         <TopHeader
-          appLogo={'/images/headerLogo.svg'}
+          appLogo="/images/headerLogo.svg"
           t={key => t[key]}
           headerConstants={{
             blackList: {
